app/requirements.txt
Flask
pytest
gunicorn
requests
google-cloud-storage